+++ src/client/connect/mod.rs
@@ -83,13 +83,20 @@ use std::fmt;
 
 use ::http::Extensions;
 
-#[cfg(feature = "tcp")]
-pub mod dns;
-#[cfg(feature = "tcp")]
-mod http;
-#[cfg(feature = "tcp")]
-pub use self::http::{HttpConnector, HttpInfo};
-pub use self::sealed::Connect;
+cfg_feature! {
+    #![feature = "tcp"]
+
+    pub use self::http::{HttpConnector, HttpInfo};
+
+    pub mod dns;
+    mod http;
+}
+
+cfg_feature! {
+    #![any(feature = "http1", feature = "http2")]
+
+    pub use self::sealed::Connect;
+}
 
 /// Describes a type returned by a connector.
 pub trait Connection {
@@ -260,6 +267,7 @@ where
     }
 }
 
+#[cfg(any(feature = "http1", feature = "http2"))]
 pub(super) mod sealed {
     use std::error::Error as StdError;
